moved remote player ping login into PlayerModelManager
